Move generic build functions to buildSrc
