+++ buildSrc/build.gradle.kts
@@ -1,0 +1,7 @@
+repositories {
+    jcenter()
+}
+plugins {
+    `kotlin-dsl`
+    `java-gradle-plugin`
+}
