+++ buildSrc/src/main/java/StringHelpers.kt
@@ -1,0 +1,5 @@
+/**
+ * Executes string as shell command and returns result as string.
+ */
+fun String.execute(): String = Runtime.getRuntime().exec(this)
+        .inputStream.bufferedReader().use { it.readText() }
